requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
Flask-Login
Flask-Bcrypt
python-dotenv